requirements.txt
asgiref==3.8.1
Django==5.1.1
dnspython==2.6.1
itsdangerous==2.2.0
mongoengine==0.29.1
pillow==10.4.0
pymongo==4.10.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
django-embed-video~=1.4.10